+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Explicit vshard master configuration
+- Automatic failover (switchable)
 - Unit tests
 
 ## [0.2] - 2018-10-01